change About structure
