+++ components/web-pages/About/About.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import Head from 'next/head';
+
+const About = () => {
+    return (
+        <>
+            <Head>
+                <title>Ninja List | About</title>
+                <meta name="keywords" content="ninjas" />
+            </Head>
+            <div>
+                <h1>About</h1>
+                <p>
+                    Lorem ipsum dolor sit amet, consectetur adipisicing elit.
+                    Atque delectus deserunt distinctio dolore, facere illum
+                    maxime molestias rerum tempora vitae? A alias cumque debitis
+                    eaque itaque labore, repudiandae totam vero!
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet, consectetur adipisicing elit.
+                    Atque delectus deserunt distinctio dolore, facere illum
+                    maxime molestias rerum tempora vitae? A alias cumque debitis
+                    eaque itaque labore, repudiandae totam vero!
+                </p>
+            </div>
+        </>
+    );
+};
+
+export default About;
