+++ components/web-pages/About/index.js
@@ -1,0 +1,1 @@
+export { default as About } from './About';
