+++ pages/About.js
@@ -1,30 +1,2 @@
-import React from 'react';
-import Head from 'next/head';
-
-const About = () => {
-    return (
-        <>
-            <Head>
-                <title>Ninja List | About</title>
-                <meta name="keywords" content="ninjas" />
-            </Head>
-            <div>
-                <h1>About</h1>
-                <p>
-                    Lorem ipsum dolor sit amet, consectetur adipisicing elit.
-                    Atque delectus deserunt distinctio dolore, facere illum
-                    maxime molestias rerum tempora vitae? A alias cumque debitis
-                    eaque itaque labore, repudiandae totam vero!
-                </p>
-                <p>
-                    Lorem ipsum dolor sit amet, consectetur adipisicing elit.
-                    Atque delectus deserunt distinctio dolore, facere illum
-                    maxime molestias rerum tempora vitae? A alias cumque debitis
-                    eaque itaque labore, repudiandae totam vero!
-                </p>
-            </div>
-        </>
-    );
-};
-
+import { About } from '../components/web-pages/About';
 export default About;
